+++ CHANGELOG.md
@@ -9,9 +9,14 @@
   - This overlaps with the existing `spp.get_load_forecast` method, which we want to eventually remove in favor of these two methods.
 - Add support for operating reserves
 
+### EIA
+
+- Add support to specify facets in get_dataset
+
 ### ERCOT
 
 - Added initial support for using the ERCOT API
+
 
 ## v0.24.0 - Dec 27, 2023